deduplicate utils
formerly had core/utils and a utils module, just have the module now.
