+++ autopilot/core/plots.py
@@ -36,8 +36,7 @@ pg.setConfigOptions(antialias=True)
 
 from autopilot import tasks, prefs
 from autopilot.core import styles
-from autopilot.core.utils import get_invoker
-from .utils import InvokeEvent, Invoker
+from ..utils.invoker import InvokeEvent, Invoker, get_invoker
 from autopilot.networking import Net_Node
 from autopilot.core.loggers import init_logger
