+++ autopilot/tasks/nafc.py
@@ -15,7 +15,7 @@ from autopilot.stim.sound import sounds as stim_sounds
 from autopilot.hardware import gpio
 from collections import OrderedDict as odict
 from autopilot.networking import Net_Node
-from autopilot.core.utils import find_recursive
+from autopilot.utils import find_recursive
 
 from autopilot import prefs
 import pdb
