--- autopilot/utils/update_pilot.sh
@@ -1,4 +1,0 @@
-killall python
-cd ~/git/RPilot
-git pull
-/usr/rpilot/launch_pilot.sh
