+++ autopilot/viz/trial_viewer.py
@@ -8,9 +8,10 @@ Warning:
 # renders a standalone webpage with bokeh of trial data for all subjects in the data folder
 import sys
 import os
-sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 
-from autopilot.core import utils
+import autopilot.utils
+
+sys.path.append(os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))))
 
 from glob import glob
 import argparse
@@ -31,7 +32,7 @@ def load_subject_data(data_dir, subject_name, steps=True, grad=True):
 
     # pilot_db_fn = [fn for fn in os.listdir(data_dir) if fn == 'pilot_db.json'][0]
     # pilot_db_fn = os.path.join(data_dir, pilot_db_fn)
-    pilot_db = utils.load_pilotdb(reverse=True)
+    pilot_db = autopilot.utils.load_pilotdb(reverse=True)
 
     # find pilot for subject
     pilot_name = pilot_db[subject_name]
@@ -200,7 +201,7 @@ if __name__ == '__main__':
         data_dir = args.dir
 
     # TODO Make arg
-    active_mice = utils.list_subjects()
+    active_mice = autopilot.utils.list_subjects()
